+++ bc/utils/wagtail_hooks.py
@@ -1,6 +1,4 @@
-from django.templatetags.static import static
 from django.urls import path, reverse
-from django.utils.html import format_html
 
 from wagtail.admin.menu import AdminOnlyMenuItem
 from wagtail.admin.rich_text.converters.html_to_contentstate import BlockElementHandler
@@ -148,10 +146,3 @@ def register_big_text_feature(features):
     }
 
     features.register_converter_rule("contentstate", feature_name, db_conversion)
-
-
-@hooks.register("insert_editor_css")
-def editor_css():
-    return format_html(
-        '<link rel="stylesheet" href="{}">', static("bc_admin_ui/editor.css")
-    )